requirements.txt
batchgenerators==0.25
einops==0.7.0
matplotlib==3.4.3
ml_collections==0.1.1
monai==1.3.0
nibabel==5.2.0
nni==3.0
numpy==1.24.1
opencv_python==4.8.1.78
Pillow==9.3.0
Pillow==10.2.0
pydensecrf==1.0rc3
scipy==1.11.4
SimpleITK==2.2.1
SimpleITK==2.3.1
skimage==0.0
Surface_Distance_Based_Measures==0.1
tifffile==2023.7.10
timm==0.9.12
torch==2.0.1+cu118
torchtoolbox==0.1.8.2
torchvision==0.15.2+cu118
tqdm==4.65.0
